add order types to wallet package, make price undefined if order type is market
