+++ libs/wallet/src/index.ts
@@ -4,3 +4,4 @@ export * from './hooks';
 export * from './connect-dialog';
 export * from './connectors';
 export * from './storage-keys';
+export * from './types';
